Fix almost all warnings

+++ doomgeneric/r_segs.c
@@ -395,7 +395,7 @@ R_StoreWallRange
     
     // calculate rw_distance for scale calculation
     rw_normalangle = curline->angle + ANG90;
-    offsetangle = abs(rw_normalangle-rw_angle1);
+    offsetangle = abs((int)rw_normalangle - rw_angle1);
     
     if (offsetangle > ANG90)
 	offsetangle = ANG90;
@@ -403,7 +403,7 @@ R_StoreWallRange
     distangle = ANG90 - offsetangle;
     hyp = R_PointToDist (curline->v1->x, curline->v1->y);
     sineval = finesine[distangle>>ANGLETOFINESHIFT];
-    rw_distance = FixedMul (hyp, sineval);
+    rw_distance = FixedMul(hyp, sineval);
 		
 	
     ds_p->x1 = rw_x = start;
